requirements.txt.txt
streamlit
PyPDF2
google-generativeai
protobuf<5,>=4.25.3
pandas
numpy
